Colocate embedding sharing ui
