+++ frontend/src/metabase/dashboard/components/DashboardHeader/DashboardHeaderButtonRow/DashboardSharingMenu/tests/setup.tsx
@@ -1,15 +1,7 @@
-/* eslint-disable i18next/no-literal-string */
-
 import userEvent from "@testing-library/user-event";
 
 import { setupEnterpriseOnlyPlugin } from "__support__/enterprise";
-import {
-  setupNotificationChannelsEndpoints,
-  setupUserRecipientsEndpoint,
-  setupUsersEndpoints,
-} from "__support__/server-mocks";
-import { setupWebhookChannelsEndpoint } from "__support__/server-mocks/channel";
-import { setupListNotificationEndpoints } from "__support__/server-mocks/notification";
+import { setupNotificationChannelsEndpoints } from "__support__/server-mocks";
 import { mockSettings } from "__support__/settings";
 import { renderWithProviders, screen } from "__support__/ui";
 import { useSelector } from "metabase/redux";
@@ -18,9 +10,7 @@ import {
   createMockDashboardState,
   createMockState,
 } from "metabase/redux/store/mocks";
-import { checkNotNull } from "metabase/utils/types";
-import Question from "metabase-lib/v1/Question";
-import type { Card, Dashboard, Notification, User } from "metabase-types/api";
+import type { Dashboard, User } from "metabase-types/api";
 import {
   createMockCard,
   createMockDashboard,
@@ -31,9 +21,7 @@ import {
 } from "metabase-types/api/mocks";
 
 import { DashboardSharingMenu } from "../DashboardSharingMenu";
-import { QuestionSharingMenu } from "../QuestionSharingMenu";
 
-// This is a fake sidebar that we can use to check if the correct redux state is getting updated
 const FakeSidebar = () => {
   const sidebar = useSelector((state) => state.dashboard.sidebar);
 
@@ -52,7 +40,6 @@ type SettingsProps = {
   isAdmin?: boolean;
   canManageSubscriptions?: boolean;
   isEnterprise?: boolean;
-  card?: Card;
   dashboardState?: Partial<DashboardState>;
 };
 
@@ -64,7 +51,6 @@ const setupState = ({
   isAdmin = false,
   canManageSubscriptions = false,
   isEnterprise = false,
-  card,
   dashboardState,
 }: SettingsProps) => {
   const tokenFeatures = createMockTokenFeatures({
@@ -90,21 +76,16 @@ const setupState = ({
     is_superuser: isAdmin,
   });
 
-  const state = createMockState({
+  return createMockState({
     settings: mockSettings(settingValues),
     currentUser: {
       ...user,
       permissions: {
         can_access_subscription: canManageSubscriptions,
       },
-      qb: {
-        card,
-      },
     } as User,
     dashboard: createMockDashboardState(dashboardState),
   });
-
-  return state;
 };
 
 export function setupDashboardSharingMenu({
@@ -153,63 +134,6 @@ export function setupDashboardSharingMenu({
   renderWithProviders(
     <div>
       <DashboardSharingMenu dashboard={dashboard} />
-      <FakeSidebar />
-    </div>,
-    { storeInitialState: state },
-  );
-}
-
-export function setupQuestionSharingMenu({
-  isPublicSharingEnabled = false,
-  isEmbeddingEnabled = false,
-  isEmailSetup = false,
-  isSlackSetup = false,
-  isAdmin = false,
-  canManageSubscriptions = false,
-  isEnterprise = false,
-  hasPublicLink = false,
-  question: questionOverrides = {},
-  alerts = [],
-}: {
-  question?: Partial<Card>;
-  hasPublicLink?: boolean;
-  alerts?: Notification[];
-} & SettingsProps) {
-  const card = createMockCard({
-    name: "My Cool Question",
-    public_uuid: hasPublicLink && isPublicSharingEnabled ? "1337bad801" : null,
-    ...questionOverrides,
-  });
-
-  const state = setupState({
-    isPublicSharingEnabled,
-    isEmbeddingEnabled,
-    isEmailSetup,
-    isSlackSetup,
-    isAdmin,
-    canManageSubscriptions,
-    isEnterprise,
-    card,
-  });
-
-  const user = checkNotNull(state.currentUser);
-
-  setupListNotificationEndpoints({ card_id: card.id }, alerts);
-  setupUsersEndpoints([user]);
-  setupUserRecipientsEndpoint({
-    users: [user],
-  });
-  setupWebhookChannelsEndpoint();
-
-  if (isEnterprise) {
-    setupEnterpriseOnlyPlugin("audit_app");
-    setupEnterpriseOnlyPlugin("application_permissions");
-    setupEnterpriseOnlyPlugin("collections");
-  }
-
-  renderWithProviders(
-    <div>
-      <QuestionSharingMenu question={new Question(card)} />
       <FakeSidebar />
     </div>,
     { storeInitialState: state },
