+++ frontend/src/metabase/documents/components/DocumentHeader.tsx
@@ -27,11 +27,11 @@ import {
 import { isWithinIframe } from "metabase/utils/iframe";
 import type { Document } from "metabase-types/api";
 
-import { DocumentPublicLinkPopover } from "../../embedding/components/PublicLinkPopover";
 import { trackDocumentPrint } from "../analytics";
 import { DOCUMENT_TITLE_MAX_LENGTH } from "../constants";
 import { getListCommentsQuery } from "../utils/api";
 
+import { DocumentPublicLinkPopover } from "./DocumentHeader/DocumentPublicLinkPopover/DocumentPublicLinkPopover";
 import S from "./DocumentHeader.module.css";
 
 const saveButtonTransition: TransitionProps["transition"] = {
