--- frontend/src/metabase/embedding/components/PublicLinkPopover/index.ts
@@ -1,3 +1,0 @@
-export { QuestionPublicLinkPopover } from "./QuestionPublicLinkPopover";
-export { DashboardPublicLinkPopover } from "./DashboardPublicLinkPopover";
-export { DocumentPublicLinkPopover } from "./DocumentPublicLinkPopover";
