--- frontend/src/metabase/embedding/components/SharingMenu/index.ts
@@ -1,2 +1,0 @@
-export * from "./DashboardSharingMenu";
-export * from "./QuestionSharingMenu";
